requirements.txt
nomic
numpy
scikit-learn
matplotlib
pandas
requests
python-dotenv
openai
playwright